changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11809 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Custom Hardware/KATRIN Hardware/eGun/OReGunModel.h
@@ -79,9 +79,9 @@
 - (ORObjectProxy*) y220Object;
 - (BOOL) moving;
 - (void) setMoving:(BOOL)aMoving;
-- (unsigned)currentTrackIndex;
-- (unsigned)validTrackCount;
-- (NSPoint) track:(unsigned)i;
+- (NSUInteger)currentTrackIndex;
+- (NSUInteger)validTrackCount;
+- (NSPoint) track:(NSUInteger)i;
 
 - (BOOL) absMotion;
 - (void) setAbsMotion:(BOOL)aAbsMotion;
